Fix DuckDB compatibility: vectorized NULL handling, unnamed params, BLOB to TEXT, and test updates

+++ tests/testthat/test-database.R
@@ -256,6 +256,6 @@ test_that("get_study", {
 
 test_that("get_nrows", {
   db <- create_test_db()
-  expect_vector(get_nrows(db), integer(), length(sensors))
+  expect_vector(get_nrows(db), numeric(), length(sensors))
   cleanup_test_db(db)
 })
